requirements.txt
grobid-client-python>=0.0.7
pymupdf>=1.23.0
beautifulsoup4>=4.12.0

sentence-transformers>=2.2.2

transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0

chromadb>=0.4.15

groq>=0.3.0 

streamlit>=1.28.0


python-dotenv>=1.0.0
pyyaml>=6.0.1
tiktoken>=0.5.0